+++ WrmOriginCard.vue
@@ -10,7 +10,7 @@
       style="width: 48px; max-height: 48px; padding-right: 10px"
       class="wrm-flex-item"
       @error="imageError">
-    <div class="wrm-flex-item-grow">
+    <div class="wrm-flex-item-grow wrm-ellipsis">
       <div style="font-size: 14px">
         <div>
           <slot name="task"></slot>
@@ -21,7 +21,7 @@
         style="justify-content: flex-start; font-weight: bold">
         <i class="fas fa-lock wrm-flex-item wrm-green"
           style="padding-right: 2px"></i>
-        <span v-if="domain">
+        <span v-if="domain" class="wrm-ellipsis">
           <span class="wrm-green">https</span><span class="wrm-dark-gray">://{{domain}}</span>
         </span>
         <span v-else>